requirements.txt
Django
mysql-connector-python
PyMySQL
pyserial